requirements.txt
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
docutils==0.21.2
fpdf==1.7.2
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
joblib==1.4.2
Kivy==2.3.0
Kivy-Garden==0.1.5
nltk==3.8.1
numpy==2.0.1
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
pyparsing==3.1.2
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
pytesseract==0.3.10
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
textblob==0.18.0.post0
tqdm==4.66.5
uritemplate==4.1.1
urllib3==2.2.2
